correction du modifySauce dans /controller/sauce.js

+++ backend/controllers/sauces.js
@@ -40,23 +40,41 @@ exports.modifySauce = (req, res, next) => {
             imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
         }
         : { ...req.body };
-    if (req.file) {
-        Sauce.findOne({ _id: req.params.id })
-            .then((sauce) => {
+    delete sauceObject._userId;
+
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            // si l'utilisateur n'est pas à l'origine du post de la sauce
+            if (req.auth.userId !== sauce.userId) {
+                res.status(403).json({ message: `Non autorisé !` })
+            }
+            // si l'utilisateur est à l'origine du post sauce, update de la sauce
+            else {
                 const filename = sauce.imageUrl.split("images/")[1];
-                // suppression de l'image de la sauce qui sera remplacée
-                fs.unlink(`images/${filename}`, (error) => {
-                    if (error) console.log(error);
-                });
-            })
-            .catch((error) => res.status(500).json({ error }));
-    }
-    Sauce.updateOne(
-        { _id: req.params.id },
-        { ...sauceObject, _id: req.params.id }
-    )
-        .then(() => res.status(200).json({ message: "Objet modifié !" }))
-        .catch((error) => res.status(400).json({ error }));
+                // si présence d'un fichier
+                if (req.file) {
+                    // suppression de l'image
+                    fs.unlink(`images/${filename}`, (error) => {
+                        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                            .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+                            .catch(error => res.status(401).json({ error }));
+                    });
+                }
+                // si non présence d'un fichiern update 
+                else {
+                    Sauce.updateOne(
+                        { _id: req.params.id },
+                        { ...sauceObject, _id: req.params.id }
+                    )
+                        .then(() => res.status(200).json({ message: "Objet modifié !" }))
+                        .catch((error) => res.status(400).json({ error }));
+
+                }
+            };
+        })
+        .catch((error) => res.status(500).json({ error }));
+
+
 };
 
 // controlleur route DELETE => Supprime une sauce
@@ -64,7 +82,7 @@ exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
             if (sauce.userId != req.auth.userId) {
-                res.status(401).json({ message: 'Not authorized' });
+                res.status(403).json({ message: 'Non autorisé' });
             } else {
                 const filename = sauce.imageUrl.split('/images/')[1];
                 // suppr du fichier avec fonction unlink du package file system (fs)
